casts: drop redundant and const-discarding casts
v7_data_end re-cast its own filsys_edition_t *fs parameter; p7fs_write_inode
widened (uint16_t)uid back to uint32_t for no effect.  mount.filsys.c freed
its strdup'd packing/arch/geom.byteorder through (void *) casts that hid a
const qualifier on owned memory.

filsys_geom_t.byteorder is caller-owned, so type it char * and have
fsck.filsys.c own its detected copy too, then free both -g and detected
byteorder after filsys_apply_geom consumes them.

+++ pdp7fs.c
@@ -327,7 +327,7 @@ int p7fs_write_inode(p7fs_t *fs, uint32_t ino, const p7_inode_t *ip) {
     d[0] = ip->mode & P7_MAXWORD;
     for (int i = 0; i < P7_NIADDR; i++)
         d[1 + i] = ip->addr[i] & P7_MAXWORD;
-    d[8]  = (uint32_t)(uint16_t)ip->uid & P7_MAXWORD;
+    d[8]  = (uint16_t)ip->uid & P7_MAXWORD;
     d[9]  = (uint32_t)(-(int32_t)ip->nlink) & P7_MAXWORD;   /* stored negative */
     d[10] = ((ip->size + 1) / 2) & P7_MAXWORD;              /* bytes -> words (round up) */
     /* d[11] (uniq) is left untouched */
